+++ optaplanner-examples/src/main/java/org/optaplanner/examples/tsp/domain/location/Location.java
@@ -75,7 +75,7 @@ public abstract class Location extends AbstractPersistable {
      * @param location never null
      * @return a positive number, the distance multiplied by 1000 to avoid floating point arithmetic rounding errors
      */
-    public abstract int getDistance(Location location);
+    public abstract long getDistance(Location location);
 
     public double getAirDistanceDouble(Location location) {
         // Implementation specified by TSPLIB http://www2.iwr.uni-heidelberg.de/groups/comopt/software/TSPLIB95/
